Respect "and" operator from match query when using only_subwords
A match query with operator "and" involving a decompounded word (e.g. "viel kunststoff") was interpreted as "+viel +(kunst stoff)". This is most likely not the desired result, as "kunst" and "stoff" are not synonyms but both have to occur. This commit changes this behavior to be "+viel +kunst +stoff". The case only_subwords=false is left untouched because it is not clear what the desired result is in that case.